fix(background): close the session-mirror review follow-ups
